Fix type error

+++ benchmarks/filtsmooth.py
@@ -143,7 +143,7 @@ class DenseGridOperations:
 
         self.locations = np.arange(0.0, info["tmax"], step=info["dt"])
         self.dense_locations = np.random.uniform(
-            low=0.0, high=1.2 * info["tmax"], size=1.2 * len(self.locations)
+            low=0.0, high=1.2 * info["tmax"], size=int(1.2 * len(self.locations))
         )
 
         _, self.observations = statespace.generate_samples(
